# swagger.json
{
  "x-generator": "NSwag v13.15.9.0 (NJsonSchema v10.6.8.0 (Newtonsoft.Json v9.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
    "/api/Library": {
      "get": {
        "tags": [
          "Library"
        ],
        "operationId": "Library_GetBooksByTitle",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookListing"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Library/lend/{userId}/{bookId}": {
      "post": {
        "tags": [
          "Library"
        ],
        "operationId": "Library_LendBook",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Library/return/{bookId}": {
      "post": {
        "tags": [
          "Library"
        ],
        "operationId": "Library_ReturnBook",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BookListing": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "books": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Book"
            }
          },
          "reservations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
      "Book": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "listing": {
            "$ref": "#/components/schemas/BookListing"
          },
          "loanInfo": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/LoanInfo"
              }
            ]
          }
        }
      },
      "LoanInfo": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "loanedDate": {
            "type": "string",
            "format": "date"
          },
          "loanedUser": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "User": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "borrowedBooks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Book"
            }
          },
          "reservedBooks": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "loanedDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        }
      }
    }
  }
}